Managing apache2.conf on eggplant in Puppet.
We want to minimize configuration in the global file like /etc/apache2/apache2.conf, and instead much prefer fragment configurations in /etc/apache2/conf.d. But the property I'm going to change (KeepAliveTimeout) is in apache2.conf, and so I have no other choice.

Perhaps this same apache2.conf can be shared by other hosts, but that needs more careful research and analysis into how other hosts use Apache. I'm not yet ready to spend that much time on this, especially when this is a legacy repository.

So I'm just going to create eggplant local module to get this done. Please feel free to refactor.