Reduce keep-alive timeout.
Stats indicate that during the US day time, Apache on eggplant is saturated. All 150 processes are forked and they are completely busy. Looking at mod_status output, majority of them are in the keep-alive wait mode.

I'm keeping the keep-alive option, so that when a browser hits the page, it can request all the static resources and stuff in one TCP connection. But I'm reducing the timeout so that once a wave of requests are over, apache moves on to serve another connection.

This setting is pre-tested on the production and significantly reduced
the apache worker count and improved the response time of Wiki.