Reduced I/O timeout.
Apache worker process is precious resources, and we often see the requests saturate all 150 worker processes. So getting them process requests quickly is important.

To that end, don't try too hard to wait on unresponsive clients/backends. It's better to give up on one request and moves on to another, than trying to stick to a request that's not getting anywhere.